Improve rendering. Remove debug code. Slight refactor
