+++ packages/tinacms/src/toolkit/fields/plugins/streamliners/discussion-plugin/index.ts
@@ -6,11 +6,9 @@ export * from './types';
 export * from './plugins/discussion-adapter';
 export * from './plugins/discussion-plugin';
 export * from './utils/annotations-store';
-export * from './utils/comment-annotations';
 export * from './hooks/use-annotation-state';
 export * from '../suggestion-plugin/suggestion-plugin';
 export * from '../suggestion-plugin/components/suggestion-leaf';
 export * from '../suggestion-plugin/components/suggestion-toolbar-button';
 export * from '../suggestion-plugin/components/suggestion-card';
-export * from '../suggestion-plugin/utils/suggestion-state-synchronizer';
 export * from '../suggestion-plugin/utils/suggestion-helpers';
