--- packages/tinacms/src/toolkit/fields/plugins/streamliners/discussion-plugin/utils/annotation-id.ts
@@ -1,6 +1,0 @@
-export const createAnnotationId = (prefix: string) => {
-  if (typeof crypto !== 'undefined' && typeof crypto.randomUUID === 'function') {
-    return crypto.randomUUID();
-  }
-  return `${prefix}-${Date.now()}-${Math.floor(Math.random() * 1_000_000)}`;
-};
