+++ packages/tinacms/src/toolkit/fields/plugins/streamliners/suggestion-plugin/index.ts
@@ -1,4 +1,3 @@
 export * from './suggestion-plugin';
 export * from './components/suggestion-leaf';
 export * from './components/suggestion-toolbar-button';
-export * from './utils/suggestion-state-synchronizer';
